+++ great_expectations/core/util.py
@@ -364,27 +364,6 @@ def ensure_json_serializable(data):
         )
 
 
-def safe_deep_copy(data, memo=None):
-    """
-    This method makes a copy of a dictionary, applying deep copy to attribute values, except for non-pickleable objects.
-    """
-    if isinstance(data, (pd.Series, pd.DataFrame)) or (
-        pyspark and isinstance(data, pyspark.sql.DataFrame)
-    ):
-        return data
-
-    if isinstance(data, (list, tuple)):
-        return [safe_deep_copy(data=element, memo=memo) for element in data]
-
-    if isinstance(data, dict):
-        return {
-            key: safe_deep_copy(data=value, memo=memo) for key, value in data.items()
-        }
-
-    # noinspection PyArgumentList
-    return copy.deepcopy(data, memo)
-
-
 def requires_lossy_conversion(d):
     return d - decimal.Context(prec=sys.float_info.dig).create_decimal(d) != 0
